Update patch set 5
Patch Set 5:

All overlays are OK.
But consider this -
There is an overlay to change the value of a bool to true.
But in the original file too the value is true. In that case the overlay is not required.
If you have time go through overlays and delete ones which fall into the above category.

Patch-set: 5